+++ assets/javascripts/kitten/components/meters/spinner/text.js
@@ -1,0 +1,30 @@
+import React from 'react'
+import renderer from 'react-test-renderer'
+import { Spinner } from './index'
+import COLORS from '../../../constants/colors-config'
+
+describe('<Spinner />', () => {
+  let component
+
+  describe('simple Spinner ', () => {
+    beforeEach(() => {
+      component = renderer.create(<Spinner />).toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('Spinner with multiple props', () => {
+    beforeEach(() => {
+      component = renderer
+        .create(<Spinner color={COLORS.primary1} value={50} />)
+        .toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+})